style: move tera::endpoint in tera::middleware
